[core] Cleanup picker tests
