+++ docs/pages/x/api/date-pickers/desktop-date-range-picker.json
@@ -117,10 +117,7 @@
   },
   "name": "DesktopDateRangePicker",
   "styles": { "classes": [], "globalClasses": {}, "name": "MuiDesktopDateRangePicker" },
-  "spread": false,
-  "forwardsRefTo": "HTMLDivElement",
   "filename": "/packages/x-date-pickers-pro/src/DesktopDateRangePicker/DesktopDateRangePicker.tsx",
-  "inheritance": null,
   "demos": "<ul><li><a href=\"/x/react-date-pickers/date-range-picker/\">Date Range Picker </a></li>\n<li><a href=\"/x/react-date-pickers/validation/\">Validation</a></li></ul>",
   "packages": [
     { "packageName": "@mui/x-date-pickers-pro", "componentName": "DesktopDateRangePicker" }
