+++ docs/pages/x/api/date-pickers/mobile-time-picker.json
@@ -111,7 +111,6 @@
   "name": "MobileTimePicker",
   "styles": { "classes": [], "globalClasses": {}, "name": "MuiMobileTimePicker" },
   "filename": "/packages/x-date-pickers/src/MobileTimePicker/MobileTimePicker.tsx",
-  "inheritance": null,
   "demos": "<ul><li><a href=\"/x/react-date-pickers/time-picker/\">Time Picker</a></li>\n<li><a href=\"/x/react-date-pickers/validation/\">Validation</a></li></ul>",
   "packages": [
     { "packageName": "@mui/x-date-pickers-pro", "componentName": "MobileTimePicker" },
