+++ packages/x-date-pickers/src/DateTimePicker/tests/DateTimePicker.test.tsx
@@ -3,17 +3,9 @@ import { DateTimePicker } from '@mui/x-date-pickers/DateTimePicker';
 import { screen } from '@mui/monorepo/test/utils/createRenderer';
 import { expect } from 'chai';
 import { createPickerRenderer, stubMatchMedia } from 'test/utils/pickers-utils';
-import { describeValidation } from '@mui/x-date-pickers/tests/describeValidation';
 
 describe('<DateTimePicker />', () => {
-  const { render, clock } = createPickerRenderer({ clock: 'fake' });
-
-  describeValidation(DateTimePicker, () => ({
-    render,
-    clock,
-    views: ['year', 'month', 'day', 'hours', 'minutes'],
-    componentFamily: 'picker',
-  }));
+  const { render } = createPickerRenderer();
 
   it('should render in mobile mode when `useMediaQuery` returns `false`', () => {
     const originalMatchMedia = window.matchMedia;
