+++ packages/x-date-pickers/src/MobileDateTimePicker/tests/describes.MobileDateTimePicker.test.tsx
@@ -11,12 +11,15 @@ import {
   getTextbox,
 } from 'test/utils/pickers-utils';
 import { MobileDateTimePicker } from '@mui/x-date-pickers/MobileDateTimePicker';
+import { describePicker } from '@mui/x-date-pickers/tests/describePicker';
 
 describe('<MobileDateTimePicker /> - Describes', () => {
   const { render, clock } = createPickerRenderer({
     clock: 'fake',
     clockConfig: new Date(2018, 2, 12, 8, 16, 0),
   });
+
+  describePicker(MobileDateTimePicker, { render, fieldType: 'single-input', variant: 'mobile' });
 
   describeValidation(MobileDateTimePicker, () => ({
     render,
