+++ packages/x-date-pickers/src/MobileTimePicker/tests/describes.MobileTimePicker.test.tsx
@@ -18,12 +18,15 @@ import {
   getTextbox,
 } from 'test/utils/pickers-utils';
 import { MobileTimePicker } from '@mui/x-date-pickers/MobileTimePicker';
+import { describePicker } from '@mui/x-date-pickers/tests/describePicker';
 
 describe('<MobileTimePicker /> - Describes', () => {
   const { render, clock } = createPickerRenderer({
     clock: 'fake',
     clockConfig: new Date(2018, 2, 12, 8, 16, 0),
   });
+
+  describePicker(MobileTimePicker, { render, fieldType: 'single-input', variant: 'mobile' });
 
   describeValidation(MobileTimePicker, () => ({
     render,
