+++ packages/x-date-pickers/src/StaticDatePicker/tests/StaticDatePickerKeyboard.test.tsx
