+++ packages/x-date-pickers/src/tests/describePicker/index.ts
@@ -1,0 +1,1 @@
+export { describePicker } from './describePicker';
